Fix inconsistency in pagination example
The 'Next' links in the pagination examples do not match the depicted
code block.
According to the code block they should be using &rarr; but in the
examples &raquo; is used.
While this is still a matter of taste, at least the previous and next
links should be kept consistent.

It's a pitty that IE7 does not support generated content, so we cannot
move the symbols from the markup to the stylesheet in a :before or
:after pseudo element (yet), as I would consider them presentation
instead of content.