Changing pagination's boundary items
Previously you would use 'prev' and 'next to round the corners of the
items in the pagination list.  I changed it to use the 'first-child'
and 'last-child' becuase you want the corners rounded regardless of
whether or not you have a list item marked with the 'prev' or 'next'
class.  For example, I have the following pagination list:

    [1][...][12][13][14][15][16][...][23]

I could have added the 'prev' and 'next' classes to the boundary
items (1 and 23), however it's semantically incorrect, and I though
my solution was more elegant.