Fixed size reset for collapses in IE (and other browsers which does not support transitions).
reset() is called after the collapsible content has been resized by a
transition, to set the width or height to auto.

However, when transitions are not supported, the reset() callback is called
immediately, which causes the reset to take place BEFORE the width/height change
it is supposed to reset.
This makes the reset() call useless, since the content gets the static
width/height anyways.

This changes uses setTimeout to make sure that the reset() callback is called
after the width/height is set.