Update the makefile to preserve the JavaScript copyright when using YUI Compressor
The project I'm working on uses uncompressed resources during development, but compressed resources for release, courtesy of the YUI Compressor.  We are also building Bootstrap from source so that we can change variables.less as necessary.

When we compress the bootstrap/js/bootstrap.js produced by the makefile, it doesn't have the copyright prepended to bootstrap.min.js, and that copyright is stripped by the YUI Compressor as it does not begin with "/*!".

This simple change updates the copyright to begin with "/*!" and prepends it to both the compressed and uncompressed JavaScript files.