promote using data-toggle over rel for tooltip
data-toggle is consistently used throughout Bootstrap for
toggling UI related content. The rel attribute is also
intended for expressing the relationship between the current
URL and the URL being linked to.

In addition (and the initial motiviation for this change),
the HTML5 validator [1] marks the HTML as invalid for using
unregistered rel values.

[1] http://validator.w3.org/