added dismiss option for closing tooltips onclick
This is actually most useful for popovers which inherit
from Tooltip, so the change could be moved there if it
is less useful on Tooltip, but figured someone might
see the value on Tooltip as well.

There are cases where Popovers in my UI have a link
in the content. The user is unable to reach the link
before the popover disappears. Adding a delay merely
turns it into a game to see just how quickly the user
can make up their mind and click the link before it
disappears. Hence the need for a dismiss option.

The current code only support a click dismiss option
as that seemed to be the most likely action. If the
user clicks outside the popover, it goes away.