Add "sibling" placement option
prefixing placement with "sibling" will insert the tooltip *before* the source element, and position it based on the source element's position relative to its offset parent.

this change allows for tooltips that will remain properly positioned on window resize & content change, without inserting the tooltip "in" the element, which would be invalid for <input> source elements, for example.