ensure definitions in mixins.less do not output
this change allows one to...

    @import "twitter/bootstrap/mixins.less"

multiple times, without any output that is not explicitly called for.

to accomplish this,

1. three remaining mixins (.clearfix, .hide-text and .input-block-level) have been made parametric
2. non-parametric versions (simple calls to the parametric) have been added to forms.less and utilities.less, in order to allow these to be used as class attribute values and to maintain backward compatibility
3. all bootstrap-internal calls make use of the parametric versions, to ensure nothing breaks.

an example of this, in fact, working may be seen in the changes to the generated bootstrap.css and bootstrap-responsive.css.  the former still includes the output rulesets, while bootstrap-responsive.css no longer includes those definitions (which would be duplication).