Docs 2.1- add class 'fade' to myModal
The description says modal will fade and slide, but was missing 'fade' class as in 2.0.4 so it was just appearing.
Obviously not the place for suggestions, but I didn't want to noise up the issues list. It might be nice to have two classes, fade and slide.  Fade would transition the opacity, and slide would transition the 'top'.  Since some people might be interested in only the subtler fade effect, without going for the more drastic fade/slide combo.