Tooltip Works for FF on D3 rect
The bootstrap-tooltip.js was not working when applied to a graph created
using the D3 library. On further inspection I discovered that the
getPositions method was returning width:undefined.

This patch adds a check on the this.$element[0].offsetWidth and
this.$element[0].offsetHeight to make sure they are defined. If they are
not then they default to 0;

Change-Id: I71af2b37b4bf171707712711f0ccc2ad7f8c289d