Update js/bootstrap-collapse.js
Added a check to the COLLAPSIBLE DATA-API function to add the .collapsed
class to all collapsed elements in an accordion group instead of just the
element that has been click. This is so as to be able to style collapsed
elements differently to open elements. e.g. showing an up or down arrow
on the data-toggle=collapse element based on the .collapsed class.