Adding an 'expanded' class to click target
To help with styling elements that are targets for the collapse plugin, we required a class when the element was in the expanded state. Because the default state was collapsed, there was no difference in class on the target element from the default state through the first click to expand.