Add experimental off-canvas layout
Based on:
https://github.com/bradfrost/this-is-responsive/blob/gh-pages/patterns/layout-offcanvas-right.html

It lacks a nice transition effect due to a strange bug happening
only in Chrome at certain viewport sizes.

Other gotchas: only works with a sidebar coming in from the right,
and the required javascript is ugly and written in the HTML file
because I don't feel comfortable enough to figure out how to write
a noob-level plugin for Bootstrap.