Prefer to collapse siblings
This fix is to prefer to collapse siblings instead of every target. The situation in which this is useful is when two collapsible navbars exist on one screen. The reasoning for the if block is simple, if we have a sibling, prefer that, if not, then do it the 'legacy' way, as I do not know what other dependencies their may be.