Add rem as default unit for font-sizes, including fallback
This commit adds a simple mixin which calculates a pixel based
value for font-sizes based on a provided rem-value.

Including the mixin:

    .rem-font-size(@size) // while @size is a unit-less number

The `@size`-variable is based on the root-font-size – which is 10px
in this case.

This commit also changes all font-sizes from px-based values to the
usage of the rem-size fallback.