Hide tooltip if parent element moved
No mouseout event is triggered when the parent element for the tooltip
changes position, so the tooltip doesn't dissapear even if the cursor
isn't hover it anymore.

A check every 500 milliseconds has been added that looks for the
position of the parent item and if it changed hides the tooltip.