Make title and content functions cachable.
This commit introduces an opt-in flag for tooltips and popovers which
allows the widgets to cache the return values from the `title` and
`content` functions, where they have been provided. This is useful for
when the title and content need computation, but are not going to
change. Obviously the functions themselves can be memoised, but this
approach also avoids the cost of repeated (and pointless) function
calls.