Added repositioning mechanism for dynamic content.
Tooltip and popover content may have cause to grow dynamically, and its
final form cannot always be determined synchronously. An important use
case for this is popover elements that fetch their content via ajax.
This commit adds a mechanism for such dynamic elements to be reflown
when their content changes. This must still be managed manually by the
caller, but at least it is now possible without external modification. A
typical system for doing this is expected to be calling the `reposition`
method on the tooltip/popover when the content is changed and its
position needs to be recalculated.