Remove unnecessary code
The only point I can see to this code is to fire the 'slid' event 
the moment pause is called, instead of when it finishes sliding.

But I had to add the '&& !this.sliding' check because if 
transition.end was called while sliding, the content being slid 
away vanishes. So after fixing that, the rest of this code seems 
pointless.