Pull For Issue #7384 Typeahead Positioning
A Typeahead control in the page footer or on the right hand side of the
page can result in overflow issues when the menu is shown.

Allow the specification of left (default) or right alignment with the
input control.
Allow the specification of display below (default) or above the input
control.