Proposed fix for allowing 'show' and 'hide' events in Collapse be cancelled before completion
Added tracking for if preventDefault has been called similar to how
'transitioning' is currently tracked.  If that has occured, the
elements are returned their previous states in show() and hide()
respectively.  The more complicated use case is for when a group of
elements is being managed by a parent element, and calling show()
triggers the other elements to be hidden.  I changed the plugin
defintion to take an argument (a boolean) to pass to hide() in this
(and only this) case so that reseting this.defaultPrevented only
happens in the orginating show() function.  The other way to go here so
that the plugin definition is not changed is to call a different hide
function (hideOpenSiblings()?) in show() to manage the other elements
that never reset this.defaultPrevented to 0.