UglifyJS v2 no longer supports the `-nc` flag.
In UglifyJS v1, `-nc` causes initial comments in input code to be
discarded.  This is used to strip the per-file copyright comments, so
a global one can be prepended.

UglifyJS v2 has POLA-violated to some extent by not only removing this
flag but by also _changing the default behaviour_ such that this is
now the default and preserving copyrights must be specified using the
`--comments` option.