More distinguished state of toggleable buttons
Use darker colors for .btn.active (i.e. pressed) state when button is
set toggleable via data-toggle="button...". Also after button is pressed
it becomes brighter on :hover, :focus, :active.

So when button is in the released state hovering over makes it darker
as default, but after clicking on it the color becomes even darker and
it easy to see that the button is pressed. On the other side, hovering
over the pressed button makes it brighter and clicking on it brightens
button further.

This not only helps out to see easier the current button state
but also shows which way the state will change.