e.preventDefault() on tooltip/popover toggle
When the tooltip is created on an anchor element with an href the default event is not prevented. This is annoying as i have quite a lot of popovers as progressive enhancement but if no javascript is present, the user is sent to the correct page. Now I am doing another 'click' listener which prevents the default behaviour and then fires the popover (much like the twitter bootrap demo page does) but I guess it would be nice to have that behaviour directly in the bootstrap js.

#6500 was submitted some time ago but was rejected for not working with the -wip branches