Add border calc for @btn-default-border: darken(@btn-default-bg, 20%)
The 20% should keep the #ccc border on the current #fff bg.

This change sets up easier for people to extend default button color - still they would have to specify a new bg and decide if they want a 20% or 5% darken on borders (other btn styles are 5%), but at least groundwork is there for uniform calculations.

That said, I cut-pasted the change in far less time than I wrote this so... it's never too big an issue. Relates to 01c46bf.