Added a like style to span elements
Like in pagination, when the item is 'active' someone would try to use a span instead of an anchor tag (because there's no point on having an href in links do page refreshes). This pull tryies to solve the missing styles for span elements on navigations so one can change the markup of anchor to span without losing the styling.