Removed parentheses in declaration of mixins (in mixins.less) which is going to be accessible as a class. Removed the classes which has already been declared as mixins without parentheses, as they will be added to the final CSS.
In LESS any class can be used as a mixin, but when we declare a mixin and afterwards add a class with the same name, LESS actually duplicates the mixin. When we apply e.g. the .clearfix to the .row-selector, the clearfix CSS is added twice to the final CSS.