Fix calculated width on media-body
On some browsers, the media width is calculated using only the media-body. I don't know why. So adding a media-object within a pull-left node, pushes the media-body out of the viewable area. This fixes that problem. It only happens when there isn't a fixed with. If the media object can expand to a set width it's fine, such as the the example usage: http://getbootstrap.com/components/#media. However some browsers are unable to put the media node in a variable width node such as a popover.

I made a jsfiddle demonstrating the problem. Check it out in firefox. Toggle the media-body css setting. http://jsfiddle.net/givingstage/BFds2/