Added list feedback classes
Adds support for list feedback, styled after the form feedback classes. These classes allow styling each element of a list to enhance feedback. This is appropriate in cases where there a list of constraints a user must meet: a list can be styled as a "checklist", and updated as requirements are fulfilled.

Usage is very simple:
<ul class="list-feedback">
    <li>This item is has no feedback class.</li>
    <li class="list-success">Item with list-success</li>
    <li class="list-warning">Item with list-warning</li>
    <li class="list-error">Item with list-error</li>
</ul>