Add option called 'remotePlacement' for modals
In default, when option 'remote' is passed to modal, element with the
selector '.modal-content' (according to documentation
http://getbootstrap.com/javascript/#modals ) will be replaced with
gathered remote content. In some cases we don't want to lose
header/footer section of the modal, which have to be inside
'.modal-content' element (stylesheets rules are applied for section
header/footer inside '.modal-content').

This option may decrease the number of duplicated code:
We can image that, page can have one default template for modals
(including HTML structure and stylesheets for header/footer) and only
content is different. We don't want to have header/footer structure
in every remote content source, so the header/footer should be kept.