Update variables.less
I have been frustrated with the fact that Bootstrap does not have table captions integrated into the code. Being concerned with usability and accessibility, and the fact that around 75% of my clients use captions on their tables, we really need to add this feature. I will also add to the tables.less file as well.