Fixes arrows in tooltips wrongly centred at times
Tooltips (and popovers) showing arrows wrongly positioned at times (centred, more commonly seen when zoom is enabled in the browser). The reason was, for example, with an arrow to be shown at the bottom of the tooltip, `getViewportAdjustedDelta()` returned `delta` with `left == 0` and the test `delta.left` failed. In these cases, when it happened *and* there was an adjustment to be made (height != actualHeight), the arrow ended up being vertically centred instead of at the bottom of the tooltip.