Accept DOM nodes as the tooltip / popover content.
When a DOM node is passed to an HTML tooltip, that node is only moved if
it is not already in the tooltip. Otherwise, `empty()` is used instead of
`detach()` before appending the `title` to avoid memory leaks.
If a DOM node is passed to a plain text tooltip then its text is copied.