Use only the input-group email field in the example
It seems strange that the original example has *two* "enter email"
fields - one with an sr-only label, and one with an input-group-addon.
If nothing else, this blows up the inline form as it's too wide.
Logically, it makes no sense as an example form (compared to all other
form examples, that have a single email field). If the intent was to
also show the use of input-group with the extra addon, then this change
makes the most sense to me...