Add basic example for collapsible components
Currently the "collapse" documentation jumps straight into the
accordion example and then mentions in passing the slightly simpler
example of using a single component to expand/collapse a region after
that.

This leaves some room for misunderstandings; for example I spent a
while not realising that I'd set data-target in my <a> links, which
makes the page jump around when clicked as the default action isn't
prevented (see [1]).

This proposes a simple example first up, using a link and a button to
expand a div.  It also explicitly mentions use of href and
data-target.  The more involved accordion example then follows
unmodified.

[1] https://github.com/twbs/bootstrap/issues/2001