Update collapse example with link example
As a follow-on to 9928144dc4e8301ee0dd02ac073682d789b2adb1, update the
collapse example with another demonstration of using a link to control
the region.  This specifically calls-out the href v data-target
arguments and also shows the default-expanded example just for
reference.

With this the page is getting a bit long, so cut down the lorem ipsum
text from the accordian sample html (not the live example; where the
full text is left to pad it out).

Also remove the now obsolete note on using a single button and add a
note on the prevention of default-action to the usage section.