Modify collapse.js unit test
The unit test would execute `hide` on an already hidden element, and
then check if it has height set (which is added by the transition). This
only worked previously, because collapse.js erroniuosly executed
`toggle` which then trigged `show` instead of `hide`. Now that `hide` is
actually executed, no transition happens, so the height value is not set
by the animation. Hence in order to test the `hide` animation, the
element must first be shown.