Improve tooltip container placement
Tooltips are currently getting appened to all matching container
selectors. This changes it to use jquery's `closest` selector to
traverse up the dom tree until the container is found.

This fixes issue #16443 by limiting the containers found to 1 so when
the tooltip is removed, excess dom nodes aren't left in the tree.

[Fixes #16443]