Fix for accordion-like collapse toggle functionality
At the moment only child elements (of element referenced with `data-parent="#accordion"`) having `.panel` class toggle properly.
Since this is not emphasized or mentioned in docs I believe it's a bug. Otherwise it is very restricting design wise as `.panel` class has to be used. And due to it's own css styling it isn't fit for every situation where toggle functionality could be used.

This fix makes it more flexible and allows proper accordion-like toggling functionality for any child element (of any class).