Add JavaScript configuration for Hound
This change uses Bootstrap's existing `js/.jshintrc` file
to configure Hound's hosted JSHint instance.

It leaves existing JavaScript linting as-is
because Hound does not support ESLint or JSCS yet.

On each pull request to Bootstrap,
Hound will comment on any JavaScript style violations in-line,
like this:

![screenshot](https://images.thoughtbot.com/hound/javascript-example.png)

If you update the pull request to adopt a suggestion,
the comment will be hidden.

It leaves the existing linting done by Grunt + Travis.

Hound is free for open source projects
and is open source itself:

https://github.com/thoughtbot/hound