Fix #17964 (CSP Violation Errors)
Some browsers are lazy in updating elements after transitions. An effective workaround is to call an element's property, which causes the browser to resync (reflow). However, the existing implementation invokes the Function generator, which violates Content Security Policy (where active) and crashes the application. The proposed change is actually a step back towards v3, but it works as intended.