v4a1 docs Add hints to resp utils test cases display
The displayed output to show the effects of the various classes
    `.hidden-{xs,sm,md,lg,xl}-{up,down}`
is really rather cryptic and hard to understand.  Even more so, it is
hard to relate the output with which classes were active to select the
particular on/off status boxes.

This change adds a `title="..."` text to show on hover which class or
combination of classes has caused a particular status display box
to be active.  This may help ease the confused mind a little bit.

One thought was to add the same text inside the displayed status boxes,
but that looked rather ugly and obtrusive.  Hence using `title="..."`

Note that the four cases of classes `.hidden-{sm,md,lg,xl}-only` did
not have titles added.  These are docs-only classes defined within the
/docs/scss/ files, ending up in the docs.min.css file.  Not bootstrap
classes.  Whether the need to be created for the docs test case example
might indicate these are needed by bootstrap is nonobvious to me.