v4 docs Restore fancy balloons to dashboard example
Due to problems with v3 demo paths (#16699) change 079c35f6134 temporarily
substituted dummy images for the previous colored balloons. That problem
has been fixed now in v4 so this change re-enables the balloons in the
example.

This is not a critical need, but rather is done so that comparisons of
v4-dev and v3 don't look so jarring.