Allow multiple dropdown button addons after inputs
In an `input-group`, it is now possible to add multiple button addons after the
input with dropdown menus. These have to be placed in separate `input-group-btn`
elements because otherwise the JavaScript breaks. Focus state now highlights the
complete group of controls too, instead of only the input.

Because of CSS limitations, highlighting the entire group only happens when the
input is the first element in the group. This is because the siblings before the
input cannot be selected and `:focus` does not bubble.

Also, the documentation now features a couple of examples with multiple buttons
as addons to inputs. One example, with plain buttons (no dropdowns) was already
working, but not being shown, so I still think it is useful to have it.