Use bootstrap-toc plugin for sidebar navigation
Previously, the sidebar navigation for every page was constructed by
hand, which means each file needs to be kept up-to-date with the
headings on the respective page. This commit removes these files to
instead inject the navigation items automatically, using the Bootstrap
TOC plugin.

http://afeld.github.io/bootstrap-toc/

Use of this plugin results in (nearly) identical markup and identical
functionality of the side nav bar, with a lot less clutter in the
repository.