Extend form docs with example form errors
Forms should show a error message whenever there is an error. This can
be done nicely in bootstrap with the `.text-help` class. But until know
people hat to figure that out by themselves.

This commit extends the example html with dummy error messages for the
warning and danger cases. It also shows a case where a text-help
paragraph is combined with an text-muted paragraph, wich requires an
additional `m-b-0` at the moment (see
https://github.com/twbs/bootstrap/issues/18703).
In addition to that I extended the introduction to specify the
`.text-help` class as the right way to show error messages in general.
The `.text-help` is not explained in the form-docs yet, but it should.