Button toggling - trigger change event on input
Bootstrap’s .button styles can be applied to other elements, such as labels, to provide checkbox or radio style button toggling.

And when the checkbox or radio state is changed, there should be triggered
the change event. Currently, the change event is triggered on the Button, which is not correct. Only input fields do support the change event.