Line height should be the same as .form-control
When $line-height is changed the .form-control line height is increased, see https://github.com/twbs/bootstrap/blob/v4-dev/scss/_forms.scss#L12. But the line height of .input-group-addon was hard coded so it would not increase in size. The text .input-group-addon will not center vertical that way.

This change fixes that issue.