Mention contributors in copyright
I am not a lawyer but adding Contributors to copyright because Twitter Inc isn't the only entity owning the copyright. The copyright is collectively owned by all contributors, or more precisely: every contributor owns a copyright on the code they have contributed.

The only way to retain full copyright over code is to ask for signing contributor's agreement before merging in changes. This is, in fact, what some companies who license free software commercially do.

As I mentioned, I am not an expert in this domain. Any opinion from people who are more well-versed with this subject (Copyright and Intellectual Property Rights) would be of great help

If this has already been discussed, then I apologize for bringing it up again as I couldn't find it via the issue tracker