Adding distribution CSS file in "main".
In some cases developers just need to pull in the distribution CSS file inside their projects just with a <link> tag.  But if we want to take advantage of resources like : https://www.npmjs.com/package/wiredep, and pull in the CSS automatically, we can't because that file is missing in the list of "main"  files.