Add auto to `m-` spacer utility
margin-*: auto is an invaluable tool when using flexbox positioning,
and with bootstrap 4 now supporting flexbox, it would be good
to have it built in as a base utility to avoid the need for messy
one-off classes added by end users.

This adds auto as a spacer value to use with the spacing utility.
.m-t-auto is an example of the class being used. With this new class,
the explicit .m-x-auto class is no longer needed as it is generated
like the others.

Documentation updated with note adding `auto`