minor update to the .card-img class
when testing this on my own the image don't take the full width but also has some extra space (padding on the .card itself that could be removed to make the card better). Your site documented it the way it should look like but the CSS doesn't work towards that.