Fix IE11 table layout issue with flexbox
When applying flexbox to a table in IE, the table cells have strange widths. This is because IE11 sets TD to display: table-cell even when the parent has display: flex. The solution is to manual set the TDs to display: block. There are also wrapping issues that mean table and tbody also need display: block. That bug limits our ability to be more specific by saying "tr.row > td".