Add border width to input height variables
This commit adds the border width (top + bottom) to the input height
variables by default. This makes IMHO much more sense because every
element is set to `box-sizing: border-box`, so the border height should
count to the total element height.

This commit introduces some handy helper functions to convert `rem`
units to `px` and vice versa. This is required to calculate the total
input height because the border width is defined in `px` and the other
height related variables (`padding`, `font-size` etc.) are defined in
`rem`.