Update docs for IE10+IE11 support of auto margins
IE10 and IE11 do not properly support the use of margin-auto on flex parents with a justify-content value other than flex-start. This PR removes the examples showcasing those two classes at play together in favor of simpler ones with _just_ the margin-auto utilities.

More details can be found at https://stackoverflow.com/a/37535548, which has been linked to from the docs as well.

Closes #24319 and closes #24413.